# lib/api-zod/src/generated/types/passwordResetInput.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * Multi-Tenant SaaS POS System API — Phase 1 (Foundation)
 * OpenAPI spec version: 0.1.0
 */

export interface PasswordResetInput {
  /** @minLength 8 */
  newPassword: string;
}