# lib/api-zod/src/generated/types/userInput.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * Multi-Tenant SaaS POS System API — Phase 1 (Foundation)
 * OpenAPI spec version: 0.1.0
 */

export interface UserInput {
  /** @minLength 3 */
  username: string;
  /** @minLength 8 */
  password: string;
  /** @minLength 1 */
  fullName: string;
  roleId: string;
  phone?: string;
  email?: string;
  isActive?: boolean;
}